+++ lib/src/screens/home_screen.dart
@@ -24,7 +24,7 @@ const _telegramUrl = 'https://t.me/ivan_it_net';
 const _vkUrl = 'https://vk.com/ivan_yurievich_it';
 const _donateUrl = 'https://dzen.ru/ivanyurievich?donate=true';
 const _supportEmail = 'ai@ivan-it.net';
-const _appVersion = '1.0.9';
+const _appVersion = '1.0.10';
 
 enum _AppLanguage {
   ru('ru'),